Moved files around to use the new vite; refactor file extensions to .jsx;
