+++ src/components/Admin/AddEmployee.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import { Component } from "react";
 // import { API_URL } from "../../config";
 import "./AddEmployee.css";
