+++ src/components/Admin/DeleteEmployee.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const DeleteEmployee = () => (
   <div>
     <h1>this is delete employee page</h1>
