+++ src/components/Admin/EditEmployee.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import { Component } from "react";
 // import { API_URL } from "../../config";
 import "./EditEmployee.css";
