+++ src/components/Card.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Card.css";
 
 const Card = ({
