+++ src/components/CardList2.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Card from "./Card";
 
 const CardList2 = ({ movies, requestIndiCard }) => {
