+++ src/components/Hamburger.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Hamburger.css";
 
 const Hamburger = () => {
