+++ src/components/SearchBox.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./SearchBox.css";
 
 const SearchBox = ({ searchChange }) => {
