+++ src/components/ToggleButton.jsx
@@ -1,4 +1,4 @@
-import React, { useRef } from "react";
+import { useRef } from "react";
 import "./ToggleButton.css";
 
 const ToggleButton = ({ handleClick }) => {
