--- src/config.js
@@ -1,1 +1,0 @@
-// export const API_URL = `${process.env.REACT_APP_BACKEND_URL}/api/achs-dir`
